mmc: fix 32bit silent truncation of 64bit offset in mmc_get_env_addr()
The function mmc_get_env_addr() return the address (offset) of the
enviroment variables. Internally it calculates a 64bit value, but
the value is silently truncated by the provided 32bit return value.
For offsets greater than 4GB the environment is read from and written
to the wrong place in e. g. the flash memory and may override
arbitrary data.

+++ board/nxp/common/sdhc_boot.c
@@ -16,7 +16,7 @@
 
 #define ESDHC_DEFAULT_ENVADDR	0x400
 
-int mmc_get_env_addr(struct mmc *mmc, int copy, u32 *env_addr)
+int mmc_get_env_addr(struct mmc *mmc, int copy, u64 *env_addr)
 {
 	u8 *tmp_buf;
 	u32 blklen, code_offset, code_len, n;
